Quick and dirty bugfix.
Issue: When attempting to train via the XMI database, the message
"The table de.julielab.jcore.types.Token does not exist." appears.
That's true because we didn't say in which schema to look and the
activate data schema in the respective (local) costosys.xml was wrong.
So for now I just add the "_data_xmi" prefix which is some kind of
convention. Not sure how this should be handled better.